+++ src/main/kotlin/com/bftcom/library/book/repository/BookRepository.kt
@@ -6,8 +6,8 @@ interface BookRepository {
 
     fun create(book: Book): Book
     fun getAll(): List<Book>
-    fun getById(id: Int): Book?
-    fun update(id: Int, book: Book): Book
-    fun delete(id: Int)
+    fun getById(id: Long): Book?
+    fun update(id: Long, book: Book): Book
+    fun delete(id: Long)
 
 }
